sudarshan-backend/requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0

# ── Auth / crypto ────────────────────────────────────────────────────────────
pyjwt>=2.8,<3.0
bcrypt>=4.0,<5.0
cryptography>=42,<45

# ── Config ───────────────────────────────────────────────────────────────────
pydantic[email]>=2.7,<3.0
pydantic-settings>=2.0,<3.0

# ── WebSockets ───────────────────────────────────────────────────────────────
websockets>=12,<15

# ── Logging / dev utilities ──────────────────────────────────────────────────
python-dotenv>=1.0,<2.0

# ── Stream API ───────────────────────────────────────────────────────────────
stream-chat>=4.0,<5.0